amplify/functions/process-bedrock/requirements.txt
boto3>=1.34.131
pypdf>=3.17.0
pdf2image
openpyxl>=3.1.2
python-pptx>=0.6.23

markitdown[all]
